+++ src/entities/post/api/index.ts
@@ -1,1 +1,2 @@
 //
+export { addPost, editPost } from "./post.api";
